Fix react-select to not disappear available options from the dropdown
The latest version of React-Select includes a slew of changes to how the data is managed. One of these changes was resulting in spurious calls to the backend data store, which was causing the cached and filtered display options to be overwritten with empty and with repeated values of the already selected value. The simple solution exists in the main webapp branch. For now, this branch includes changes to eliminate a ton of React warnings about duplicated items and one change to correctly handle the cache.