Fixing bug in react-select where the selected values couldn't be deleted in one click
Was a bug with the fix to the warnings about duplicate field keys. Fixed so that options and values will always match keys, but the initial loading values, which have no labels, will get their own key. This avoids the duplicates warning and retains the ability of react-select to remove selected values.